[quant] Add nonuniform quantization mapping functions
[ghstack-poisoned]
